clean up types and work on firebase integration of data
