--- client/src/pages/notablePerson/dummyData.ts
@@ -1,42 +1,0 @@
-export const data = {
-  notablePersonId: 1001,
-  notablePersonName: 'Michael Jackson',
-  notablePersonPictureUrl: 'http://lorempixel.com/150/150/',
-  notablePersonLabels: [
-    'Musician',
-    'Non-Political',
-    "Jehova's Witness",
-  ],
-  notablePersonEvents: [
-    {
-      eventId: 100,
-      eventQuote: '"If you enter this world knowing you are loved and you leave this world knowing the same, then everything that happens in between can be dealt with."',
-      eventSource: 'http://tmz.io/mjackson/bigevent',
-      eventSourceName: 'TMZ',
-      userComment: 'A very inspirational quote by MJ!! It summarizes his world view pretty clearly.',
-      userAvatar: 'http://lorempixel.com/20/20/',
-      postedBy: 'Mr Cookie',
-      postedAt: 1494999526,
-    },
-    {
-      eventId: 101,
-      eventQuote: '"In a world filled with hate, we must still dare to hope. In a world filled with anger, we must still dare to comfort. In a world filled with despair, we must still dare to dream. And in a world filled with distrust, we must still dare to believe."',
-      eventSource: 'http://politico.io/mj/speech',
-      eventSourceName: 'Politico',
-      userComment: 'This was published yesterday, he was quoted in the award ceremony.. Maybe this will shed some light on his beliefs!',
-      userAvatar: 'http://lorempixel.com/20/20/',
-      postedBy: 'Awesome Gal',
-      postedAt: 1495001889,
-    },
-    {
-      eventId: 102,
-      eventQuote: '"I am going to search for my star until I find it. It is hidden in the drawer of innocence, wrapped in a scarf of wonder."',
-      eventSource: 'http://nyt.co/news/rt',
-      eventSourceName: 'New York Times',
-      userComment: 'Does this mean he believes in astrology? Maybe not. I will post another article on this.',
-      userAvatar: 'http://lorempixel.com/20/20/',
-      postedBy: 'Kale Chips',
-      postedAt: 1495001897,
-    },
-  ],
-}
