+++ client/src/pages/notablePerson/notablePerson.tsx
@@ -9,33 +9,17 @@ import {pick} from '../../utils/utils'
 import {Events} from './events'
 import {common} from '../../common.styles'
 import {styles} from './notablePerson.styles'
-import {data} from './dummyData'
-
-export interface PilotData {
-  notablePersonId: number,
-  notablePersonName: string,
-  notablePersonPictureUrl: string,
-  notablePersonLabels: string[],
-  notablePersonEvents: {
-    eventId: number,
-    eventQuote: string,
-    eventSource: string,
-    eventSourceName: string,
-    userComment: string,
-    userAvatar: string,
-    postedBy: string,
-    postedAt: number,
-  }[]
-}
+import {PilotData} from '../../../../typings/dataSchema'
 
 interface IProps {
   pilotData: PilotData | {}
 }
 
 function mapStateToProps(state: State): IProps {
-  return pick(state, [
-    'pilotData',
-  ])
+  console.log(state.pilotData)
+  return {
+    pilotData: state.pilotData,
+  }
 }
 
 const actionCreators = pick(actions, [
@@ -50,19 +34,23 @@ class NotablePersonClass extends React.Component<ComponentProps, undefined> {
     const {props: p} = this
     p.requestPilotData()
   }
+  componentDidUpdate() {
+    const {props: p} = this
+    console.log(p.pilotData)
+  }
   render() {
 
     return (
       <div className={css(common.page)}>
         <div className={css(styles.notablePersonTitleContainer)}>
-          <img className={css(styles.notablePersonPhoto)} src={data.notablePersonPictureUrl}/>
+          {/*<img className={css(styles.notablePersonPhoto)} src={data.notablePersonPictureUrl}/>*/}
           <div className={css(styles.notablePersonText)}>
             <h1 className={css(styles.notablePersonTitle)}>Religion, politics, and ideas of...</h1>
-            <h2 className={css(common.titleTypography, styles.notablePersonName)}>{data.notablePersonName}</h2>
+            {/*<h2 className={css(common.titleTypography, styles.notablePersonName)}>{data.notablePersonName}</h2>*/}
             {/*{this.renderLabels()}*/}
           </div>
         </div>
-        <Events data={data}/>
+        {/*<Events data={data}/>*/}
       </div>
     )
   }
