+++ client/src/redux/selectors.ts
@@ -30,9 +30,3 @@ export function showGlobalSpinner(state: State) {
 export function hasResults(state: State) {
   return state.searchResults && state.searchResults.hits && state.searchResults.hits.length > 0 || false
 }
-
-/*
-export function getPilotData(state: State) {
-  return state.setPilotData
-}
-*/
